Add multi_transformers
