+++ src/azure-cli/azure/cli/command_modules/keyvault/_validators.py
@@ -6,7 +6,6 @@
 import argparse
 import base64
 import binascii
-from collections.abc import Iterable
 from datetime import datetime
 import re
 
@@ -375,22 +374,3 @@ def validate_storage_disabled_attribute(attr_arg_name, attr_type):
         attr_arg = attr_type(enabled=(not disabled))
         setattr(ns, attr_arg_name, attr_arg)
     return _validate
-
-
-def _extract_subresource_name_from_single_output(output, id_parameter):
-    if not getattr(output, id_parameter):
-        resource_name = None
-    else:
-        items = getattr(output, id_parameter).split('/')
-        resource_name = items[4] if len(items) > 4 else None
-
-    setattr(output, 'name', resource_name)
-    return output
-
-
-def extract_subresource_name(id_parameter='id'):
-    def _extract_subresource_name(output):
-        if isinstance(output, Iterable):
-            return [_extract_subresource_name_from_single_output(item, id_parameter) for item in output]
-        return _extract_subresource_name_from_single_output(output, id_parameter)
-    return _extract_subresource_name
