+++ src/azure-cli/azure/cli/command_modules/keyvault/commands.py
@@ -11,10 +11,11 @@ from ._client_factory import (
     keyvault_client_vaults_factory, keyvault_client_private_endpoint_connections_factory,
     keyvault_client_private_link_resources_factory, keyvault_data_plane_factory)
 
-from ._transformers import filter_out_managed_resources
+from ._transformers import (
+    extract_subresource_name, filter_out_managed_resources,
+    multi_transformers)
 
 from ._validators import (
-    extract_subresource_name,
     process_secret_set_namespace, process_certificate_cancel_namespace,
     validate_private_endpoint_connection_id)
 
@@ -104,7 +105,9 @@ def load_command_table(self, _):
 
     # Data Plane Commands
     with self.command_group('keyvault key', kv_data_sdk) as g:
-        g.keyvault_command('list', 'get_keys', transform=[filter_out_managed_resources, extract_subresource_name(id_parameter='kid')])
+        g.keyvault_command('list', 'get_keys',
+                           transform=multi_transformers(
+                               filter_out_managed_resources, extract_subresource_name(id_parameter='kid')))
         g.keyvault_command('list-versions', 'get_key_versions', transform=extract_subresource_name(id_parameter='kid'))
         g.keyvault_command('list-deleted', 'get_deleted_keys', transform=extract_subresource_name(id_parameter='kid'))
         g.keyvault_custom('create', 'create_key', doc_string_source=data_doc_string.format('create_key'))
@@ -120,7 +123,8 @@ def load_command_table(self, _):
         g.keyvault_custom('download', 'download_key')
 
     with self.command_group('keyvault secret', kv_data_sdk) as g:
-        g.keyvault_command('list', 'get_secrets', transform=[filter_out_managed_resources, extract_subresource_name()])
+        g.keyvault_command('list', 'get_secrets',
+                           transform=multi_transformers(filter_out_managed_resources, extract_subresource_name()))
         g.keyvault_command('list-versions', 'get_secret_versions', transform=extract_subresource_name())
         g.keyvault_command('list-deleted', 'get_deleted_secrets', transform=extract_subresource_name())
         g.keyvault_command('set', 'set_secret', validator=process_secret_set_namespace,
